+++ vmcnet/train/vmc.py
@@ -1,9 +1,6 @@
 """Main VMC loop."""
-import collections
 import logging
-import os
 from typing import Callable, Dict, Tuple, TypeVar
-from vmcnet.utils.checkpoint import RunningEnergyVariance, RunningMetric
 
 import jax
 import jax.numpy as jnp
@@ -16,54 +13,6 @@ import vmcnet.utils as utils
 D = TypeVar("D")
 P = TypeVar("P")  # represents a pytree or pytree-like object containing model params
 O = TypeVar("O")  # represents optimizer state
-
-
-def make_metropolis_step(
-    proposal_fn: Callable[[P, D, jnp.ndarray], Tuple[D, jnp.ndarray]],
-    acceptance_fn: Callable[[P, D, D], jnp.ndarray],
-    update_data_fn: Callable[[D, D, jnp.ndarray], D],
-) -> Callable[[P, D, jnp.ndarray], Tuple[jnp.float32, D, jnp.ndarray]]:
-    """Factory to create a function which takes a single metropolis step.
-
-    Following Metropolis-Hastings Markov Chain Monte Carlo, a transition from one data
-    state to another is split into proposal and acceptance. When used in a Metropolis
-    routine to approximate a stationary distribution P, the proposal and acceptance
-    functions should satisfy detailed balance, i.e.,
-
-        proposal_prob_ij * acceptance_ij * P_i = proposal_prob_ji * acceptance_ji * P_j,
-
-    where proposal_prob_ij is the likelihood of proposing the transition from state i to
-    state j, acceptance_ij is the likelihood of accepting a transition from state i
-    to state j, and P_i is the probability of being in state i.
-
-    Args:
-        proposal_fn (Callable): proposal function which produces new proposed data. Has
-            the signature (params, data, key) -> proposed_data, key
-        acceptance_fn (Callable): acceptance function which produces a vector of numbers
-            used to create a mask for accepting the proposals. Has the signature
-            (params, data, proposed_data) -> jnp.ndarray: acceptance probabilities
-        update_data_fn (Callable): function used to update the data given the original
-            data, the proposed data, and the array mask identifying which proposals to
-            accept. Has the signature
-            (data, proposed_data, mask) -> new_data
-
-    Returns:
-        Callable: function which takes in (data, params, key) and outputs
-        (mean acceptance probability, new data, new jax PRNG key split from previous
-        one)
-    """
-
-    def metrop_step_fn(data, params, key):
-        """Take a single metropolis step."""
-        key, subkey = jax.random.split(key)
-        proposed_data, key = proposal_fn(params, data, key)
-        accept_prob = acceptance_fn(params, data, proposed_data)
-        move_mask = jax.random.uniform(subkey, shape=accept_prob.shape) < accept_prob
-        new_data = update_data_fn(data, proposed_data, move_mask)
-
-        return jnp.mean(accept_prob), new_data, key
-
-    return metrop_step_fn
 
 
 def walk_data(
@@ -125,7 +74,7 @@ def make_burning_step(
     typically don't mean much during burning) are thrown away.
 
     For more about the Metropolis step itself, see
-    :func:`~vmcnet.train.vmc.make_metropolis_step`
+    :func:`~vmcnet.mcmc.metropolis.make_metropolis_step`
     and to see it in use, see
     :func:`~vmcnet.train.vmc.vmc_loop`.
